MLBEDSW-10299: Size types and allocator fixes for ordered_map
 - Normalise the sized types used by ordered_map to present
   an interface more consistent with the standard library.
 - Set initial allocation to zero such that declaring an
   empty map allocates no storage, and add tests for the
   same. Storage is now allocated on first-use.
 - Fix potential range issue with initial hashtable size
   being greater than the chosen indexer.
 - Fix issue where it was not possible to resize up to
   the maximum indexer limit.

+++ ethosu/regor/tflite/tflite_reader.cpp
@@ -654,14 +654,14 @@ void TfLiteReader::ParseOperatorOptions(const std::shared_ptr<Operation> &operat
         case tflite::BuiltinOptions::SplitOptions:
         {
             int num_splits = GetBuiltinOptions<tflite::SplitOptions>(tflite_operator)->num_splits();
-            if ( num_splits != operation->Outputs().size() ) operation->SetPassthroughOp();
+            if ( size_t(num_splits) != operation->Outputs().size() ) operation->SetPassthroughOp();
         }
         break;
 
         case tflite::BuiltinOptions::SplitVOptions:
         {
             int num_splits = GetBuiltinOptions<tflite::SplitVOptions>(tflite_operator)->num_splits();
-            if ( num_splits != operation->Outputs().size() ) operation->SetPassthroughOp();
+            if ( size_t(num_splits) != operation->Outputs().size() ) operation->SetPassthroughOp();
         }
         break;
